refactor: restore per-package nuxt-errors.ts using kit's createErrorUtils wrapper
Each package now has its own nuxt-errors.ts that creates a buildErrorUtils
instance via the kit-wrapped createErrorUtils (which provides default
formatters), instead of importing directly from @nuxt/kit.
