+++ packages/kit/src/components.ts
@@ -3,7 +3,7 @@ import type { Component, ComponentsDir } from '@nuxt/schema'
 import { resolveModuleExportNames } from 'mlly'
 
 import { useNuxt } from './context.ts'
-import { buildErrorUtils } from './errors.ts'
+import { buildErrorUtils } from './nuxt-errors.ts'
 import * as ErrorCodes from './error-codes.ts'
 import { resolvePath } from './resolve.ts'
 import { MODE_RE } from './utils.ts'
