+++ packages/kit/src/layout.ts
@@ -2,7 +2,7 @@ import type { NuxtTemplate } from '@nuxt/schema'
 import { join, parse } from 'pathe'
 import { kebabCase } from 'scule'
 import { useNuxt } from './context.ts'
-import { buildErrorUtils } from './errors.ts'
+import { buildErrorUtils } from './nuxt-errors.ts'
 import * as ErrorCodes from './error-codes.ts'
 import { addTemplate } from './template.ts'
 import { reverseResolveAlias } from 'pathe/utils'
