+++ packages/kit/src/nuxt-errors.ts
@@ -1,0 +1,10 @@
+import { createErrorUtils } from './errors.ts'
+import { logger } from './logger.ts'
+
+export * as ErrorCodes from './error-codes.ts'
+
+export const buildErrorUtils = createErrorUtils({
+  module: 'NUXT',
+  docsBase: 'https://nuxt.com/e',
+  logger,
+})
