+++ packages/kit/src/pages.ts
@@ -2,7 +2,7 @@ import type { NuxtHooks, NuxtMiddleware } from '@nuxt/schema'
 import { defu } from 'defu'
 
 import { useNuxt } from './context.ts'
-import { buildErrorUtils } from './errors.ts'
+import { buildErrorUtils } from './nuxt-errors.ts'
 import * as ErrorCodes from './error-codes.ts'
 import type { NitroRouteConfig } from './nitro-types.ts'
 import { toArray } from './utils.ts'
