+++ packages/kit/src/plugin.ts
@@ -5,7 +5,7 @@ import type { NuxtPlugin, NuxtPluginTemplate } from '@nuxt/schema'
 import { resolveModulePath } from 'exsolve'
 import { MODE_RE, filterInPlace } from './utils.ts'
 import { tryUseNuxt, useNuxt } from './context.ts'
-import { buildErrorUtils } from './errors.ts'
+import { buildErrorUtils } from './nuxt-errors.ts'
 import * as ErrorCodes from './error-codes.ts'
 import { addTemplate } from './template.ts'
 import { resolveAlias } from './resolve.ts'
