+++ packages/nitro-server/src/nuxt-errors.ts
@@ -1,0 +1,8 @@
+import { createErrorUtils } from '@nuxt/kit'
+
+export { ErrorCodes } from '@nuxt/kit'
+
+export const buildErrorUtils = createErrorUtils({
+  module: 'NUXT',
+  docsBase: 'https://nuxt.com/e',
+})
