+++ packages/nuxt/src/compiler/plugins/keyed-functions.ts
@@ -7,7 +7,8 @@ import { camelCase } from 'scule'
 import escapeRE from 'escape-string-regexp'
 import { findStaticImports, parseStaticImport } from 'mlly'
 import { ScopeTracker, type ScopeTrackerNode, parseAndWalk, walk } from 'oxc-walker'
-import { ErrorCodes, buildErrorUtils, resolveAlias } from '@nuxt/kit'
+import { resolveAlias } from '@nuxt/kit'
+import { ErrorCodes, buildErrorUtils } from '../../core/utils/error-format.ts'
 import type { KeyedFunction } from '@nuxt/schema'
 import type { Node } from 'oxc-parser'
 import type { Import } from 'unimport'
