+++ packages/nuxt/src/components/plugins/lazy-hydration-transform.ts
@@ -2,7 +2,8 @@ import { createUnplugin } from 'unplugin'
 import MagicString from 'magic-string'
 import { camelCase, pascalCase } from 'scule'
 
-import { ErrorCodes, buildErrorUtils, tryUseNuxt } from '@nuxt/kit'
+import { tryUseNuxt } from '@nuxt/kit'
+import { ErrorCodes, buildErrorUtils } from '../../core/utils/error-format.ts'
 import { parse, walk } from 'ultrahtml'
 import { ScopeTracker, parseAndWalk } from 'oxc-walker'
 import { isVue } from '../../core/utils/index.ts'
