+++ packages/nuxt/src/components/plugins/loader.ts
@@ -4,7 +4,8 @@ import MagicString from 'magic-string'
 import { pascalCase } from 'scule'
 import { relative } from 'pathe'
 
-import { ErrorCodes, buildErrorUtils, tryUseNuxt } from '@nuxt/kit'
+import { tryUseNuxt } from '@nuxt/kit'
+import { ErrorCodes, buildErrorUtils } from '../../core/utils/error-format.ts'
 import { QUOTE_RE, SX_RE, isVue } from '../../core/utils/index.ts'
 import { installNuxtModule } from '../../core/features.ts'
 import { resolveToAlias } from '../../utils.ts'
