+++ packages/nuxt/src/components/scan.ts
@@ -2,7 +2,8 @@ import { readdir } from 'node:fs/promises'
 import { basename, dirname, extname, join, relative } from 'pathe'
 import { glob } from 'tinyglobby'
 import { kebabCase, pascalCase, splitByCase } from 'scule'
-import { ErrorCodes, buildErrorUtils, isIgnored, useNuxt } from '@nuxt/kit'
+import { isIgnored, useNuxt } from '@nuxt/kit'
+import { ErrorCodes, buildErrorUtils } from '../core/utils/error-format.ts'
 import { withTrailingSlash } from 'ufo'
 
 import { QUOTE_RE, resolveComponentNameSegments } from '../core/utils/index.ts'
