+++ packages/nuxt/src/core/cache.ts
@@ -2,7 +2,8 @@ import { mkdir, open, readFile, stat, unlink, writeFile } from 'node:fs/promises
 import type { FileHandle } from 'node:fs/promises'
 import { resolve } from 'node:path'
 import { existsSync } from 'node:fs'
-import { ErrorCodes, buildErrorUtils, createIsIgnored } from '@nuxt/kit'
+import { createIsIgnored } from '@nuxt/kit'
+import { ErrorCodes, buildErrorUtils } from './utils/error-format.ts'
 import type { Nuxt, NuxtConfig, NuxtConfigLayer } from '@nuxt/schema'
 import { logger } from '../utils.ts'
 import { hash, serialize } from 'ohash'
