+++ packages/nuxt/src/core/utils/index.ts
@@ -1,7 +1,7 @@
 export { getNameFromPath, hasSuffix, resolveComponentNameSegments } from './names.ts'
 export { getLoader, isJS, isVue, parseModuleId } from './plugins.ts'
-export { buildErrorUtils, ErrorCodes } from '@nuxt/kit'
-export type { ErrorOptions as NuxtErrorOptions } from '@nuxt/kit'
+export { buildErrorUtils, ErrorCodes } from './error-format.ts'
+export type { NuxtErrorOptions } from './error-format.ts'
 
 export function uniqueBy<T, K extends keyof T> (arr: T[], key: K) {
   if (arr.length < 2) {
