+++ packages/nuxt/src/pages/utils.ts
@@ -2,7 +2,8 @@ import { runInNewContext } from 'node:vm'
 import fs from 'node:fs'
 import { normalize } from 'pathe'
 import { joinURL } from 'ufo'
-import { ErrorCodes, buildErrorUtils, getLayerDirectories, resolveFiles, resolvePath, useNuxt } from '@nuxt/kit'
+import { getLayerDirectories, resolveFiles, resolvePath, useNuxt } from '@nuxt/kit'
+import { ErrorCodes, buildErrorUtils } from '../core/utils/error-format.ts'
 import { genArrayFromRaw, genDynamicImport, genImport, genSafeVariableName } from 'knitwork'
 import { filename } from 'pathe/utils'
 import { hash } from 'ohash'
