+++ packages/webpack/src/webpack.ts
@@ -8,7 +8,8 @@ import type { Compiler, Stats, Watching } from 'webpack'
 import { defu } from 'defu'
 import type { NuxtBuilder } from '@nuxt/schema'
 import { joinURL } from 'ufo'
-import { ErrorCodes, buildErrorUtils, logger, useNitro, useNuxt } from '@nuxt/kit'
+import { logger, useNitro, useNuxt } from '@nuxt/kit'
+import { ErrorCodes, buildErrorUtils } from './nuxt-errors.ts'
 import type { InputPluginOption } from 'rollup'
 
 import { DynamicBasePlugin } from './plugins/dynamic-base.ts'
